rename models package to entities
