+++ src/main/java/com/codurance/retropolis/controllers/CardController.java
@@ -1,8 +1,8 @@
 package com.codurance.retropolis.controllers;
 
+import com.codurance.retropolis.entities.Card;
 import com.codurance.retropolis.exceptions.CardNotFoundException;
 import com.codurance.retropolis.exceptions.ColumnNotFoundException;
-import com.codurance.retropolis.models.Card;
 import com.codurance.retropolis.requests.NewCardRequestObject;
 import com.codurance.retropolis.requests.UpVoteRequestObject;
 import com.codurance.retropolis.requests.UpdateCardRequestObject;
