+++ src/main/java/com/codurance/retropolis/services/BoardService.java
@@ -1,6 +1,6 @@
 package com.codurance.retropolis.services;
 
-import com.codurance.retropolis.models.Board;
+import com.codurance.retropolis.entities.Board;
 import com.codurance.retropolis.repositories.BoardRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
