+++ src/main/java/com/codurance/retropolis/services/UserService.java
@@ -1,7 +1,7 @@
 package com.codurance.retropolis.services;
 
+import com.codurance.retropolis.entities.User;
 import com.codurance.retropolis.exceptions.UserNotFoundException;
-import com.codurance.retropolis.models.User;
 import com.codurance.retropolis.repositories.UserRepository;
 import org.springframework.stereotype.Service;
