+++ src/test/java/com/codurance/retropolis/acceptance/cards/AddUpVoteStepDefinitionIntegrationTest.java
@@ -7,7 +7,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
 import com.codurance.retropolis.acceptance.BaseStepDefinition;
-import com.codurance.retropolis.models.Card;
+import com.codurance.retropolis.entities.Card;
 import com.codurance.retropolis.requests.UpVoteRequestObject;
 import com.codurance.retropolis.utils.HttpWrapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
