+++ src/test/java/com/codurance/retropolis/factories/CardFactoryTest.java
@@ -2,7 +2,7 @@ package com.codurance.retropolis.factories;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.codurance.retropolis.models.Card;
+import com.codurance.retropolis.entities.Card;
 import com.codurance.retropolis.requests.NewCardRequestObject;
 import org.junit.jupiter.api.Test;
